[ES|QL] Lookup join - Add new column (#225410)
## Summary
Related issue https://github.com/elastic/kibana/issues/207330

Add the possibility of adding new columns.

* Extracted addColumn and addRow to separate components.
* Make full cells clickable.
* Display warning modal if the user tries to leave with unsaved columns.
* Parse cell values to primitives before saving to ES, this is needed to
be able to create number or boolean fields.


https://github.com/user-attachments/assets/158393b9-d0e9-43be-ae45-c6f06cbf0fc5

---------

+++ src/platform/packages/private/kbn-index-editor/src/components/create_flyout.tsx
@@ -12,7 +12,7 @@ import { FormattedMessage } from '@kbn/i18n-react';
 import { toMountPoint } from '@kbn/react-kibana-mount';
 import type { FC } from 'react';
 import React, { Suspense, lazy } from 'react';
-import { distinctUntilChanged, from, skip, takeUntil } from 'rxjs';
+import { distinctUntilChanged, firstValueFrom, from, skip, takeUntil } from 'rxjs';
 import type { EditLookupIndexContentContext, FlyoutDeps } from '../types';
 
 export function createFlyout(deps: FlyoutDeps, props: EditLookupIndexContentContext) {
@@ -38,11 +38,23 @@ export function createFlyout(deps: FlyoutDeps, props: EditLookupIndexContentCont
     };
   });
 
-  const onFlyoutClose = () => {
+  const onFlyoutClose = async () => {
+    indexUpdateService.discardUnsavedChanges();
+
+    const pendingColumnsToBeSaved = await firstValueFrom(
+      indexUpdateService.pendingColumnsToBeSaved$
+    );
+
+    if (pendingColumnsToBeSaved.length) {
+      deps.indexUpdateService.setExitAttemptWithUnsavedFields(true);
+      return;
+    }
+
     props.onClose?.({
       indexName: indexUpdateService.getIndexName()!,
       indexCreatedDuringFlyout: props.doesIndexExist ? false : indexUpdateService.getIndexCreated(),
     });
+
     indexUpdateService.destroy();
     flyoutSession.close();
   };
